(refactor) 이전의 Line, Point 삭제, 리팩토링 파일 정리
