+++ src/main/java/ladder/controller/LadderController.java
@@ -2,7 +2,6 @@ package ladder.controller;
 
 import ladder.domain.*;
 import ladder.domain.rule.AlwaysDrawRule;
-import ladder.domain.rule.RandomDrawRule;
 import ladder.view.InputView;
 import ladder.view.ResultView;
 
@@ -11,13 +10,13 @@ public class LadderController {
     public void start() {
         Players players = Players.of(InputView.inputPlayerNames());
 
-        LadderGame2 ladderGame2 = LadderGame2.builder()
+        LadderGame ladderGame = LadderGame.builder()
                 .players(players)
                 .rows(InputView.inputLineCount())
                 .build();
-        Ladder2 ladder2 = ladderGame2.makeLadder(new AlwaysDrawRule());
+        Ladder ladder = ladderGame.makeLadder(new AlwaysDrawRule());
 
         ResultView.printPlayers(players);
-        ResultView.printLadder(ladder2);
+        ResultView.printLadder(ladder);
     }
 }
