+++ src/test/java/ladder/domain/LadderTest.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.TestInstance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class LadderTest {
 
@@ -15,65 +14,64 @@ class LadderTest {
 
     @BeforeAll
     public void setup() {
-        int countOfLines = 3;
-        int countOfPoints = 3;
-        ladder = Ladder.builder()
-                .countOfLines(countOfLines)
-                .countOfPoints(countOfPoints)
+        int rowNumber = 3;
+        int columnNumber = 3;
+        ladder = Ladder.builder().rowNumber(rowNumber)
+                .columnNumber(columnNumber)
                 .build();
     }
 
     @Test
-    @DisplayName("사다리 위치에 Line이 잘 생성되는지 확인")
+    @DisplayName("위치에 Line이 잘 생성되는지 확인")
     public void ladder_draw_line_test() {
         // given
-        int curLine = 0;
+        int curRow = 0;
         int curColumn = 0;
 
         // when
-        ladder.drawLine(curLine, curColumn);
+        ladder.drawLine(curRow, curColumn);
 
         // then
-        Line line = ladder.getLines().get(curLine);
-        assertThat(line.hasLine(curColumn)).isTrue();
+        Row row = ladder.getRows().get(curColumn);
+        assertThat(row.hasLine(curColumn)).isTrue();
     }
 
     @Test
-    @DisplayName("인접한 Line끼리 선이 생성되지 않는 테스트")
+    @DisplayName("같은 Row에서 인접하게 Line이 생성되지 않는 테스트")
     public void draw_adjacent_line_test() {
         // given
-        int curLine = 0;
+        int curRow = 0;
         int curColumn = 0;
         int nextColumn = 1;
-        ladder.drawLine(curLine, curColumn);
+        ladder.drawLine(curRow, curColumn);
 
         // when
-        ladder.drawLine(curLine, nextColumn);
+        ladder.drawLine(curRow, nextColumn);
 
         // then
-        Line line = ladder.getLines().get(curLine);
-        assertThat(line.hasLine(nextColumn)).isFalse();
+        Row row = ladder.getRows().get(curRow);
+        assertThat(row.hasLine(nextColumn)).isFalse();
     }
 
 
     @Test
-    @DisplayName("윗 라인에 선이 생성 되어 있는 경우 아래 라인에 선 생성안되는 테스트")
+    @DisplayName("위의 Row에 Line이 생성되어 있는 경우 아래에 Line 생성 안되는 테스트")
     public void draw_bottom_line_test() {
         // given
-        int curLine = 0;
+        int curRow = 0;
         int curColumn = 0;
-        int nextLine = 1;
+        int nextRow = 1;
         int nextColumn = 1;
-        ladder.drawLine(curLine, curColumn);
+        ladder.drawLine(curRow, curColumn);
 
         // when
-        ladder.drawLine(nextLine, curColumn);
-        ladder.drawLine(nextLine, nextColumn);
+        ladder.drawLine(nextRow, curColumn);
+        ladder.drawLine(nextRow, nextColumn);
 
         // then
-        Line line = ladder.getLines().get(nextLine);
-        assertThat(line.hasLine(curColumn)).isFalse();
-        assertThat(line.hasLine(nextColumn)).isTrue();
+        Row row = ladder.getRows().get(nextRow);
+        assertThat(row.hasLine(curColumn)).isFalse();
+        assertThat(row.hasLine(nextColumn)).isTrue();
     }
 
 }
